Add simplistic HTTP handling.
The purpose of the HTTP handlers is to make the application logic
available to the outside world.

The following routes are exposed:

    I.  GET  /
    II. POST /

A request to (I) responds with a json representation of all posts.

A request to (II) expects the request body to be of Content-Type
`application/x-www-form-urlencoded`.  The following parameters need to
be supplied:

    title:string: the title of the post to publish.  Must not be empty.
    content:string: the content of the post to publish.  Must not be empty.