Enforce uniqueness of post titles.
Post titles will be used for generating human-readable URLs (UUIDs are
not suited for this purpose).  For this use case, they need to be
unique.

An alternative strategy would be to disambiguate the URLs for posts with
the same title by adding some kind of generated string to the title.