Make structure visible.
The file structure now reflects the number and nature of the components
necessary for an event sourced system.

A convention for handling domain rules that span several individual
entities is established.  For each type domain object, a companion type,
named in the plural, is defined (if necessary).  An object of this
companion type is instantiated in the application's Init method. Compare
this to "Class" objects on the JVM or in Ruby.

For example, type `Posts` tracks the titles of all posts to enforce a
uniqueness constraint on *all* post titles.  The uniqueness constraint
is checked in `Post.publish`.  In order to do this, every post created
needs a reference to a `Posts` object.  The method `Posts.New` is
introduced to instantiate new `Post` objects, while correctly setting
the reference to the parent `Posts` object.

This is similar to the "Type Object"[1] pattern described in "Game
Programming Patterns"[2]

[1] http://gameprogrammingpatterns.com/type-object.html
[2] http://gameprogrammingpatterns.com