Streamline command handling.
The application is now a regular CommandHandler, which returns the
events generated by the domain.  It takes care of loading the
appropriate aggregate(s), dispatching the command to it and processing
the resulting events.

Commands are now defined by a `Sanitize` method, taking care of input
parameter sanitization.  Having talked to several people, the common
opinion was, that treating empty parameters as errors is part of the
business logic and thus should live in the domain object.  Due to this,
the validation errors are moved to the domain object.