Send out mails for comment authentication.
Mails are sent using `/usr/sbin/sendmail`.  This avoids having to deal
with mailing failures/queues on the application level.

The idea of `processors` is introduced.  A processor is an event handler
that works with external systems.  Processers are not invoked when
replaying events (to prevent email from being sent out again, etc).