Improve user feedback in comment workflow.
Creating comments is a process involving more than one step.  After each
step it is necessary to give the user feedback about her current
position in the process and the following step.

This commit replaces the plain redirect after creating a comment with a
page detailing explaining the next steps.  When a user clicks the link
in the email, an approval page is displayed, showing the contents of the
comment to the user again.  This is mainly because some time might have
passed between the user writing the comment and approving it.  During
this time the user might forget the actual contents of the post or
comment.