Make app run standalone.
The app can run standalone now.  Refer to `CONFIG` for configuration
settings.  TLS is supported, the path to the private key and certificate
file need to be given though.

Storing passwords in environment variables is not a good idea, because a
process's environment is available in `/proc/$pid/environ`.  The
standalone mode is provided for making development/demos easier.