Add tests for methods on main.Post.
Start implementing tests for the business logic.  This commit adds tests
for `Post.publish` and `Post.comment`.  Further tests are required at
least for `Post.authenticateComment`.

Question that arose while writing the tests: should the whole
application state be prepared in the tests or should events be
dispatched manually to the given handlers?